Fix user interface events
The JavaScript was not set up to respond to click events on the action
button correctly. I added several functions that update the text, CSS
classes, and path for POST requests.

I added a CSS class to hide and show a user dropdown menu. There is now
a function to toggle that element.

I also changed the namespace of the JavaScript Checkout object. I'll
change this a bit more in a future commit so we can test it more easily.